Fix bug in Unicode converter
A CCT to Unicode converter was introduced by #3. It was detected that
some of the subscripted iotas weren't being converted and were left as
CCT code in the output files. This was caused by the fact that Dr.
Robinson's code follows a different character precedence order standard
than the conventional beta code. This was also causing some breathings
to appear detached from their letters, particularly when the letter was
uppercase. This PR fixes this bug and updates the Unicode CSV files to
make them correct.

Example of wrong conversion of iotas: Matthew 11:23 (Hades.)

Example of wrong conversion of breathings: 1 Corinthians 1:1 (Jesus.)

The changes to the converter.py script make sure that the precedence
order is as follows (we only touched uppercase letters in this PR):

1. asterisk
2. breathing
3. accent
4. letter
5. iota subscript

Example: *(=W|

This precedence order is required by the beta-code library, which is the
library that we used in #3 to do the conversion.