Update source files and Unicode CSVs
This commit replaces the old 2005 source files
with the most recent versions and updates the
Unicode CSV files accordingly. It also adjusts
the Python converters.

@normansimonr contacted professor Maurice A.
Robinson in May, 2023, and he asked his collaborator
Jeffrey Dodson to share the latest versions of
the source files of the Byzantine text.
The files are up-to-date as of May 29, 2023
and were sent to @normansimonr by Mr. Dodson via email.

The new source files do not have the same format as the
old ones and therefore it was necessary to create
new folders for them. In consequence, the old and
new source files are not directly comparable. Should
anyone wish to compare the 2005 and the new version
of the Byzantine text, we advice comparing the
generated Unicode versions.