Replace apostrophes with right quotation marks
This commit adjusts the Unicode converter to render the beta
code ' as a right quotation mark (’) instead of as an
apostrophe. This is done with the purpose of making the
typography more consistent, as although in some text
processors both characters are rendered exactly the same,
others show them differently, and the apostrophe is
less typographically compatible with other characters than
a right quotation mark in Greek. This change is made
primarily for aesthetical reasons.