Correct edition number in TEI XML
Professor Robinson clarified that the edition of the Byzantine text
is still that of the year 2018, and that a new edition should be
explicitly named only after the ECM Mark is included in the
apparatus. The previous release updated the edition field in the
XML header to be 2023. This commit corrects it to 2018.