Add Pericope Adulterae and Acts 24 beta code
This commit adds the μ6 version of the Pericope
Adulterae (which is a variant version in the Byzantine
corpus), as well as the Byzantine alternative of Acts
24 verses 6-8. The Unicode files were already in the repository,
but the Beta code source files were not. This commit adds them.

The files were created by Dr. Maurice A. Robinson and shared with
@normansimonr by Jeffrey Dodson by Dr. Robinson's request.

These files represent RP2018.