Update B to Byz in apparatuses
The current apparatuses of the μ6 Pericope Adulterae and
the Byzantine variant of Acts 24 have the letter B
to mark minority Byzantine variants. This commit updates
the apparatuses so that the letter B is replaced by the
abbreviation Byz, in order to make the apparatuses
conform to the standard used in the rest of the New Testament.