Add Robinson's Case for Byz Priority essay
This commit adds a Markdown version of the 'The case
for Byzantine Priority' essay by Prof. Maurice Robinson.

The file has a .qmd extension because it is a Quarto file.

Quarto files are Markdown files with some extended syntax. You
can open the essay.Rproj file using RStudio (after having installed
Quarto) and use it to convert the .qmd file to Word, LaTeX, Html,
and many other formats.

The text of the essay was kindly provided by Daniel Mount as a Word
document that was subsequently converted to .qmd.